refactor: adjust city and state service into its correct package
