+++ src/main/java/com/jpcchaves/adotar/application/service/location/StateServiceImpl.java
@@ -1,12 +1,14 @@
-package com.jpcchaves.adotar.application.service.impl.v1;
+package com.jpcchaves.adotar.application.service.location;
 
 import com.jpcchaves.adotar.application.dto.state.StateDto;
-import com.jpcchaves.adotar.application.service.usecases.StateService;
+import com.jpcchaves.adotar.application.service.location.contracts.StateService;
 import com.jpcchaves.adotar.application.utils.mapper.MapperUtils;
 import com.jpcchaves.adotar.domain.exception.ResourceNotFoundException;
 import com.jpcchaves.adotar.domain.model.State;
 import com.jpcchaves.adotar.infra.repository.StateRepository;
+
 import java.util.List;
+
 import org.springframework.stereotype.Service;
 
 @Service
@@ -16,7 +18,8 @@ public class StateServiceImpl implements StateService {
   private final MapperUtils mapperUtils;
 
   public StateServiceImpl(
-      StateRepository stateRepository, MapperUtils mapperUtils) {
+      StateRepository stateRepository,
+      MapperUtils mapperUtils) {
     this.stateRepository = stateRepository;
     this.mapperUtils = mapperUtils;
   }
