+++ src/main/java/com/jpcchaves/adotar/application/service/location/contracts/StateService.java
@@ -1,6 +1,7 @@
-package com.jpcchaves.adotar.application.service.usecases;
+package com.jpcchaves.adotar.application.service.location.contracts;
 
 import com.jpcchaves.adotar.application.dto.state.StateDto;
+
 import java.util.List;
 
 public interface StateService {
