--- src/main/java/com/jpcchaves/adotar/application/service/usecases/CityService.java
@@ -1,9 +1,0 @@
-package com.jpcchaves.adotar.application.service.usecases;
-
-import com.jpcchaves.adotar.application.dto.city.CityDto;
-import java.util.List;
-
-public interface CityService {
-
-  List<CityDto> getAllCities(Long stateId, String uf);
-}
